py-service/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
idna==3.3
ifaddr==0.1.7
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
py-eureka-client==0.10.0
requests==2.26.0
spring-config-client==0.2
urllib3==1.26.7
Werkzeug==2.0.2
